rename `json` package to `json_doc`
