--- checkov/json/registry.py
@@ -1,3 +1,0 @@
-from checkov.json.base_registry import Registry
-
-registry = Registry()
