+++ checkov/json_doc/base_registry.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, Tuple
 from checkov.common.checks.base_check_registry import BaseCheckRegistry
 from checkov.common.models.enums import CheckResult
-from checkov.json.enums import BlockType
+from checkov.json_doc.enums import BlockType
 
 
 class Registry(BaseCheckRegistry):
