+++ checkov/json_doc/enums.py
