+++ checkov/json_doc/registry.py
@@ -1,0 +1,3 @@
+from checkov.json_doc.base_registry import Registry
+
+registry = Registry()
